% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{.coordinate.descent}
\alias{.coordinate.descent}
\title{Helper function for performing coordinate descent}
\usage{
.coordinate.descent(x, y)
}
\description{
Helper function for performing coordinate descent
}
