% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.iso.lin.est.R
\name{.fit.block.max.min}
\alias{.fit.block.max.min}
\title{Helper function for running block max min algorithm.
Initially implemented by Hang Deng who wrote the original block max min
paper (Hang and Zhang, 2022)}
\usage{
.fit.block.max.min(y)
}
\arguments{
\item{y}{\code{n1 x n2} numeric matrix of observed responses}
}
\description{
Helper function for running block max min algorithm.
Initially implemented by Hang Deng who wrote the original block max min
paper (Hang and Zhang, 2022)
}
