% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx.quad.est.R
\name{.fit.cvx.reg}
\alias{.fit.cvx.reg}
\title{Helper function for fitting convex (or concave) function}
\usage{
.fit.cvx.reg(x, y, L)
}
\description{
Helper function for fitting convex (or concave) function
}
