% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lin.est.R
\name{.fit.iso.reg}
\alias{.fit.iso.reg}
\title{Helper function for fitting isotonic function}
\usage{
.fit.iso.reg(x, y, L)
}
\arguments{
\item{x}{\code{n x 1} numeric vector of observed covariate vector.}

\item{y}{\code{n x 1} numeric vector of observed response vector.}

\item{L}{A scalar for the linear parameter.}
}
\description{
Helper function for fitting isotonic function
}
