% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.iso.lin.est.R
\name{.multiply.beta}
\alias{.multiply.beta}
\title{Helper function to perform dot product over list of items.}
\usage{
.multiply.beta(beta)
}
\arguments{
\item{beta}{2d vector corresponding to the parametric components}
}
\description{
Helper function to perform dot product over list of items.
}
