% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{.predict.additive.iso}
\alias{.predict.additive.iso}
\title{Helper function for predicting isotonic functions for each dimension and
adding the predicted values.}
\usage{
.predict.additive.iso(glist, x, skip.dim = NULL)
}
\description{
Helper function for predicting isotonic functions for each dimension and
adding the predicted values.
}
