% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{.predict.iso.dim}
\alias{.predict.iso.dim}
\title{Helper function for predicting isotonic for given dimension.}
\usage{
.predict.iso.dim(gs, x, skip.dim = NULL)
}
\description{
Helper function for predicting isotonic for given dimension.
}
