% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx.quad.est.R
\name{.run.cross.validation.cvx}
\alias{.run.cross.validation.cvx}
\title{Helper function for performing CV for the quadratic parameter}
\usage{
.run.cross.validation.cvx(x, y, L0 = 0)
}
\description{
Helper function for performing CV for the quadratic parameter
}
