% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lin.est.R
\name{.run.cross.validation.iso}
\alias{.run.cross.validation.iso}
\title{Helper function for performing CV for the linear parameter}
\usage{
.run.cross.validation.iso(x, y, L0 = 0)
}
\description{
Helper function for performing CV for the linear parameter
}
