% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.iso.lin.est.R
\name{.run.iterative.optimizer.block.iso}
\alias{.run.iterative.optimizer.block.iso}
\title{Helper function for running an iterative optimizer for the linear parameter}
\usage{
.run.iterative.optimizer.block.iso(x, y, L0 = 100)
}
\arguments{
\item{x}{\code{n1 x n2} numeric matrix of observed covariates. Each element
of this matrix is a list of vector. For instance, the (i,j)th element of the
matrix corresponds to x\link{i, j} = list(c(a, b)), which takes the value (a, b).}

\item{y}{\code{n1 x n2} numeric matrix of observed responses.}

\item{L0}{The upper bound of the interval of L parameters to search over
the interval is in the form of \link{-L0 x log10(n), L0 x log10(n)}.
In practice, this should be a large enough constant so the set contains
the true Lipschitz parameter L.}
}
\description{
Helper function for running an iterative optimizer for the linear parameter
}
