% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx.quad.est.R
\name{.run.optimizer.cvx}
\alias{.run.optimizer.cvx}
\title{Helper function for running optimizer for the quadratic parameter}
\usage{
.run.optimizer.cvx(x, y, L0 = 0)
}
\description{
Helper function for running optimizer for the quadratic parameter
}
