% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lin.est.R
\name{.run.optimizer.iso.cross.fit}
\alias{.run.optimizer.iso.cross.fit}
\title{Helper function for running an optimizer
for the linear parameter with crossfitting}
\usage{
.run.optimizer.iso.cross.fit(x, y, L0 = 100)
}
\arguments{
\item{x}{\code{n x 1} numeric vector of observed covariate vector.}

\item{y}{\code{n x 1} numeric vector of observed response vector.}

\item{L0}{The upper bound of the interval of L parameters to search over
the interval is in the form of \link[=-L0 \\times log10(n), L0 \\times log10(n)]{-L0 \times log10(n), L0 \times log10(n)}.
In practice, this should be a large enough constant so the set contains
the true Lipschitz parameter L.}
}
\description{
Helper function for running an optimizer
for the linear parameter with crossfitting
}
