% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lin.est.R
\name{.run.optimizer.iso}
\alias{.run.optimizer.iso}
\title{Helper function for running optimizer for the linear parameter}
\usage{
.run.optimizer.iso(x, y, L0 = 0)
}
\description{
Helper function for running optimizer for the linear parameter
}
