% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.index.est.R
\name{.run.optimizer.single.index}
\alias{.run.optimizer.single.index}
\title{Helper function for running an iterative optimizer}
\usage{
.run.optimizer.single.index(x, y, L0 = 100)
}
\arguments{
\item{x}{\code{n x d} numeric vector of the observed covariate matrix}

\item{y}{\code{n x 1} numeric vector of the observed response vector.}

\item{L0}{L0 The upper bound of the interval of L parameters to search over
the interval is in the form of \link{-L0 x log10(n), L0 x log10(n)}.
In practice, this should be a large enough constant so the set contains
the true Lipschitz parameter L.}
}
\description{
Helper function for running an iterative optimizer
}
