% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.iso.lin.est.R
\name{.split.matrix}
\alias{.split.matrix}
\title{Split two matrices with the same dimension into training/validation
in the shape of a "checker board"}
\usage{
.split.matrix(x, y)
}
\arguments{
\item{x}{\code{n1 x n2} numeric matrix of observed covariates. Each element
of this matrix is a list of vector. For instance, the (i,j)th element of the
matrix corresponds to x\link{i, j} = list(c(a, b)), which takes the value (a, b).}

\item{y}{\code{n1 x n2} numeric matrix of observed responses.}
}
\description{
Split two matrices with the same dimension into training/validation
in the shape of a "checker board"
}
