% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{.update.additive.isotonic}
\alias{.update.additive.isotonic}
\title{Helper function for fitting isotonic function across dimension.}
\usage{
.update.additive.isotonic(x, y, glist, L)
}
\description{
Helper function for fitting isotonic function across dimension.
}
