% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{.update.isotonic}
\alias{.update.isotonic}
\title{Helper function for fitting isotonic function for given dimension.}
\usage{
.update.isotonic(x, y, glist, dim, L)
}
\description{
Helper function for fitting isotonic function for given dimension.
}
