% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{.update.linear}
\alias{.update.linear}
\title{Helper function for fitting bounded linear regression.}
\usage{
.update.linear(x, y, glist, L.max = 1, L.min = 0)
}
\description{
Helper function for fitting bounded linear regression.
}
