% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lin.est.R
\name{iso.lin.est.default}
\alias{iso.lin.est.default}
\title{Title}
\usage{
\method{iso.lin.est}{default}(x, y, L = NULL, run.optim = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
