% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{predict.additive.iso.est}
\alias{predict.additive.iso.est}
\title{Inherit predict function for \code{cvx.quad.est}}
\usage{
\method{predict}{additive.iso.est}(obj, newdata = NULL)
}
\arguments{
\item{obj}{n object with S3 class \code{cvx.quad.est}.}

\item{newdata}{\code{n x d} numeric vector of new covariate matrix If not
specified, use the covariate vector from training.}
}
\value{
\code{n x 1} numeric vector of predicted values.
}
\description{
Inherit predict function for \code{cvx.quad.est}
}
\examples{
n <- 100*2; x <- matrix(runif(n*2, -1, 1), ncol=2);
y <- abs(x[,1]) - abs(x[,2]) + rnorm(n, 0, 0.1)
res.est <- additive.iso.est(x, y)
y.pred <- predict(res.est) # fitted value
y.pred <- predict(res.est, newdata=matrix(runif(n*2, -1, 1), ncol=2))
}
