% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lin.est.R
\name{predict.iso.lin.est}
\alias{predict.iso.lin.est}
\title{Title}
\usage{
\method{predict}{iso.lin.est}(obj, newdata = NULL)
}
\arguments{
\item{newdata}{}
}
\value{

}
\description{
Title
}
\examples{
n <- 100; x <- runif(n, -1, 1); y <- sin(x*4) + rnorm(n, 0, 0.1)
res.est <- iso.lin.est(x, y)
y.pred <- predict(res.est) # fitted value
y.pred <- predict(res.est, newdata=runif(n, -1, 1)) # predict new data
}
