% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.index.est.R
\name{predict.single.index.est}
\alias{predict.single.index.est}
\title{Inherit predict function for \code{single.index.est}}
\usage{
\method{predict}{single.index.est}(obj, newdata = NULL)
}
\arguments{
\item{obj}{n object with S3 class \code{single.index.est}.}

\item{newdata}{\code{n x d} numeric vector of new covariate matrix If not
specified, use the covariate vector from training.}
}
\value{
\code{n x 1} numeric vector of predicted values.
}
\description{
Inherit predict function for \code{single.index.est}
}
\examples{
n <- 100; d <- 3; x <- matrix(runif(n*d, -1, 1), nrow=n);
beta <- c(1,0,0); y <- sin(4*x\%*\%beta) + rnorm(n, 0, 0.1)
res.est <- single.index.est(x, y)
y.pred <- predict(res.est) # fitted value
y.pred <- predict(res.est, newdata=matrix(runif(n*d, -1, 1))) # predict new data
}
