% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx.quad.est.R
\name{print.cvx.quad.est}
\alias{print.cvx.quad.est}
\title{Inherit print function for \code{cvx.quad.est}}
\usage{
\method{print}{cvx.quad.est}(obj, ...)
}
\arguments{
\item{obj}{An object with S3 class \code{cvx.quad.est}.}

\item{...}{Additional control parameter}
}
\description{
Inherit print function for \code{cvx.quad.est}
}
\examples{
n <- 100; x <- runif(n, -1, 1); y <- sin(x*4) + rnorm(n, 0, 0.1)
res.est <- cvx.quad.est(x, y)
print(res.est)
}
