% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.index.est.R
\name{print.single.index.est}
\alias{print.single.index.est}
\title{Inherit print function for \code{single.index.est}}
\usage{
\method{print}{single.index.est}(obj, ...)
}
\arguments{
\item{obj}{An object with S3 class \code{single.index.est}.}

\item{...}{Additional control parameter}
}
\description{
Inherit print function for \code{single.index.est}
}
\examples{
n <- 100; d <- 3; x <- matrix(runif(n*d, -1, 1), nrow=n);
beta <- c(1,0,0); y <- sin(4*x\%*\%beta) + rnorm(n, 0, 0.1)
res.est <- single.index.est(x, y)
print(res.est)
}
