% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.index.est.R
\name{single.index.est}
\alias{single.index.est}
\title{Nonparametric LSE for single-index model with a Lipschitz link function}
\usage{
single.index.est(x, y, L0 = 100, ...)
}
\arguments{
\item{x}{\code{n x d} numeric vector of the observed covariate matrix}

\item{y}{\code{n x 1} numeric vector of the observed response vector.}

\item{L0}{L0 The upper bound of the interval of L parameters to search over
the interval is in the form of \link{-L0 x log10(n), L0 x log10(n)}.
In practice, this should be a large enough constant so the set contains
the true Lipschitz parameter L.}

\item{...}{Additional control parameters.}
}
\value{
An object with S3 class \code{single.index.est}.
}
\description{
Nonparametric LSE for single-index model with a Lipschitz link function
}
\examples{
n <- 100; d <- 3; x <- matrix(runif(n*d, -1, 1), nrow=n);
beta <- c(1,0,0); y <- sin(4*x\%*\%beta) + rnorm(n, 0, 0.1)
res.est <- single.index.est(x, y)
y.pred <- predict(res.est) # fitted value
y.pred <- predict(res.est, newdata=matrix(runif(n*d, -1, 1))) # predict new data
}
