% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx.quad.est.R
\name{cvx.quad.est.default}
\alias{cvx.quad.est.default}
\title{Title}
\usage{
\method{cvx.quad.est}{default}(x, y, L = NULL, run.optim = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
