% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx.quad.est.R
\name{cvx.quad.est}
\alias{cvx.quad.est}
\title{Title}
\usage{
cvx.quad.est(x, y, L = NULL, run.optim = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
n <- 100; x <- runif(n, -1, 1); y <- sin(x*4) + rnorm(n, 0, 0.1)
res.est <- cvx.quad.est(x, y)
y.pred <- predict(res.est) # fitted value
y.pred <- predict(res.est, newdata=runif(n, -1, 1)) # predict new data
}
