% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.iso.est.R
\name{predict.additive.iso.est}
\alias{predict.additive.iso.est}
\title{Title}
\usage{
\method{predict}{additive.iso.est}(obj, newdata = NULL)
}
\arguments{
\item{newdata}{}
}
\value{

}
\description{
Title
}
