% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.iso.lin.est.R, R/iso.lin.est.R
\name{predict.iso.lin.est}
\alias{predict.iso.lin.est}
\title{Inherit predict function for \code{iso.lin.est}}
\usage{
\method{predict}{iso.lin.est}(obj, newdata = NULL)

\method{predict}{iso.lin.est}(obj, newdata = NULL)
}
\arguments{
\item{obj}{An object with S3 class \code{iso.lin.est}.}

\item{newdata}{\code{n x 1} numeric vector of new covariate vector. If not
specified, use the covariate vector from training.}
}
\value{
\code{n x 1} numeric vector of predicted values.

\code{n x 1} numeric vector of predicted values.
}
\description{
Inherit predict function for \code{iso.lin.est}

Inherit predict function for \code{iso.lin.est}
}
\examples{
n <- 100; x <- runif(n, -1, 1); y <- sin(x*4) + rnorm(n, 0, 0.1)
res.est <- iso.lin.est(x, y)
y.pred <- predict(res.est) # fitted value
y.pred <- predict(res.est, newdata=runif(n, -1, 1)) # predict new data
n <- 100; x <- runif(n, -1, 1); y <- sin(x*4) + rnorm(n, 0, 0.1)
res.est <- iso.lin.est(x, y)
y.pred <- predict(res.est) # fitted value
y.pred <- predict(res.est, newdata=runif(n, -1, 1)) # predict new data
}
