% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx.quad.est.R
\name{print.cvx.quad.est}
\alias{print.cvx.quad.est}
\title{Title}
\usage{
\method{print}{cvx.quad.est}(obj)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Title
}
\examples{
n <- 100; x <- runif(n, -1, 1); y <- sin(x*4) + rnorm(n, 0, 0.1)
res.est <- cvx.quad.est(x, y)
print(res.est)
}
