% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lin.est.R
\name{print.iso.lin.est}
\alias{print.iso.lin.est}
\title{Inherit print function for \code{iso.lin.est}}
\usage{
\method{print}{iso.lin.est}(obj, ...)
}
\arguments{
\item{obj}{An object with S3 class \code{iso.lin.est}.}

\item{...}{Additional control parameter}
}
\description{
Inherit print function for \code{iso.lin.est}
}
\examples{
n <- 100; x <- runif(n, -1, 1); y <- sin(x*4) + rnorm(n, 0, 0.1)
res.est <- iso.lin.est(x, y)
print(res.est)
}
