% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.index.R
\name{single.index.est.default}
\alias{single.index.est.default}
\title{Nonparametric LSE for for sing-index L-lipschitz functions}
\usage{
\method{single.index.est}{default}(x, y, L0 = 100, ...)
}
\arguments{
\item{x}{\code{n x d} numeric vector of the observed covariate matrix}

\item{y}{\code{n x 1} numeric vector of the observed response vector.}

\item{L0}{L0 The upper bound of the interval of L parameters to search over
the interval is in the form of \link[=-L0 \\times log10(n), L0 \\times log10(n)]{-L0 \times log10(n), L0 \times log10(n)}.
In practice, this should be a large enough constant so the set contains
the true Lipschitz parameter L.}
}
\description{
Nonparametric LSE for for sing-index L-lipschitz functions
}
